feat: Add bbb iframe integration
