+++ app/components/public/stream/jitsi-stream.hbs
@@ -1,0 +1,1 @@
+<div id="jitsi-root" {{did-insert this.setupJitsi}} {{did-update this.setupJitsi @videoStream}} style="width: 100%; height: 100%;"></div>
