+++ app/components/public/stream/video-stream.hbs
@@ -2,5 +2,11 @@
   {{#if this.loading}}
     <div class="ui active centered inline loader"></div>
   {{/if}}
+  {{#if this.iframeUrl}}
+    <iframe src="{{this.iframeUrl}}" title="BBB" allow="camera; microphone; display-capture" allowfullscreen="true" style="width: 100%; height: 100%; border: 0px;" />
+  {{/if}}
+  {{#if (eq @videoStream.videoChannel.provider 'jitsi')}}
+    <Public::Stream::JitsiStream @videoStream={{@videoStream}} />
+  {{/if}}
 </div>
 <Public::Stream::SidePanel @event={{@event}} @videoStream={{@videoStream}} />
